requirements.txt
certifi==2017.11.5
chardet==3.0.4
click==6.7
idna==2.6
pkg-resources==0.0.0
praw==5.2.0
prawcore==0.12.0
python-dotenv==0.7.1
requests==2.18.4
update-checker==0.16
urllib3==1.22
